+++ Runtime/UnityWriter.cs
@@ -24,10 +24,24 @@ namespace SaveSystem {
         }
 
 
+        public void Write (Vector2[] vector2Array) {
+            m_writer.Write(vector2Array.Length);
+            foreach (var vector2 in vector2Array)
+                Write(vector2);
+        }
+
+
         public void Write (Vector3 vector3) {
             m_writer.Write(vector3.x);
             m_writer.Write(vector3.y);
             m_writer.Write(vector3.z);
+        }
+
+
+        public void Write (Vector3[] vector3Array) {
+            m_writer.Write(vector3Array.Length);
+            foreach (var vector3 in vector3Array)
+                Write(vector3);
         }
 
 
@@ -36,6 +50,13 @@ namespace SaveSystem {
             m_writer.Write(vector4.y);
             m_writer.Write(vector4.z);
             m_writer.Write(vector4.w);
+        }
+
+
+        public void Write (Vector4[] vector4Array) {
+            m_writer.Write(vector4Array.Length);
+            foreach (var vector4 in vector4Array)
+                Write(vector4);
         }
 
 
@@ -55,11 +76,25 @@ namespace SaveSystem {
         }
 
 
+        public void Write (Color[] colors) {
+            m_writer.Write(colors.Length);
+            foreach (var color in colors)
+                Write(color);
+        }
+
+
         public void Write (Color32 color32) {
             m_writer.Write(color32.r);
             m_writer.Write(color32.g);
             m_writer.Write(color32.b);
             m_writer.Write(color32.a);
+        }
+
+
+        public void Write (Color32[] colors32) {
+            m_writer.Write(colors32.Length);
+            foreach (var color32 in colors32)
+                Write(color32);
         }
 
 
@@ -69,23 +104,45 @@ namespace SaveSystem {
         }
 
 
-        public void Write<T> (T obj) {
-            m_writer.Write(JsonUtility.ToJson(obj));
+        public void Write (Matrix4x4[] matrices) {
+            m_writer.Write(matrices.Length);
+            foreach (var matrix in matrices)
+                Write(matrix);
         }
 
 
-        public void Write<T> (string prefabPath, T monoBehaviour) where T : MonoBehaviour {
-            if (monoBehaviour.transform.parent is not null) {
-                const string message = "MonoBehaviour for writing must be root in hierarchy";
-                throw new NotRootObjectException(message);
-            }
+        public void Write (Mesh mesh) {
+            m_writer.Write(mesh.name);
+            Write(mesh.vertices);
+            Write(mesh.uv);
+            Write(mesh.uv2);
+            Write(mesh.uv3);
+            Write(mesh.uv4);
+            Write(mesh.uv5);
+            Write(mesh.uv6);
+            Write(mesh.uv7);
+            Write(mesh.uv8);
+            Write(mesh.bounds.center);
+            Write(mesh.bounds.extents);
+            Write(mesh.bounds.max);
+            Write(mesh.bounds.min);
+            Write(mesh.bounds.size);
+            Write(mesh.colors32);
+            Write(mesh.normals);
+            Write(mesh.tangents);
+            Write(mesh.triangles);
+        }
 
-            m_writer.Write(prefabPath);
-            m_writer.Write(monoBehaviour.name);
-            m_writer.Write(monoBehaviour.enabled);
-            m_writer.Write(monoBehaviour.transform.GetSiblingIndex());
-            Write(monoBehaviour.transform.position);
-            Write(monoBehaviour.transform.rotation);
+
+        public void Write (Mesh[] meshes) {
+            m_writer.Write(meshes.Length);
+            foreach (var mesh in meshes)
+                Write(mesh);
+        }
+
+
+        public void Write<T> (T obj) {
+            m_writer.Write(JsonUtility.ToJson(obj));
         }
 
 
@@ -96,13 +153,6 @@ namespace SaveSystem {
         }
 
 
-        public void Write<T> (string prefabPath, List<T> listMonoBehaviours) where T : MonoBehaviour {
-            m_writer.Write(listMonoBehaviours.Count);
-            foreach (var monoBehaviour in listMonoBehaviours)
-                Write(prefabPath, monoBehaviour);
-        }
-
-
         public void Write<T> (T[] arrayObjects) {
             m_writer.Write(arrayObjects.Length);
             foreach (var obj in arrayObjects)
@@ -110,55 +160,104 @@ namespace SaveSystem {
         }
 
 
-        public void Write<T> (string prefabPath, T[] arrayMonoBehaviours) where T : MonoBehaviour {
-            m_writer.Write(arrayMonoBehaviours.Length);
-            foreach (var monoBehaviour in arrayMonoBehaviours)
-                Write(prefabPath, monoBehaviour);
+        public void Write (byte byteValue) {
+            m_writer.Write(byteValue);
         }
 
 
-        public void Write (byte value) {
-            m_writer.Write(value);
+        public void Write (byte[] bytes) {
+            m_writer.Write(bytes.Length);
+            foreach (var byteValue in bytes)
+                m_writer.Write(byteValue);
         }
 
 
-        public void Write (short value) {
-            m_writer.Write(value);
+        public void Write (short shortValue) {
+            m_writer.Write(shortValue);
         }
 
 
-        public void Write (int value) {
-            m_writer.Write(value);
+        public void Write (short[] shorts) {
+            m_writer.Write(shorts.Length);
+            foreach (var shortValue in shorts)
+                m_writer.Write(shortValue);
         }
 
 
-        public void Write (long value) {
-            m_writer.Write(value);
+        public void Write (int intValue) {
+            m_writer.Write(intValue);
         }
 
 
-        public void Write (char value) {
-            m_writer.Write(value);
+        public void Write (int[] ints) {
+            m_writer.Write(ints.Length);
+            foreach (var intValue in ints)
+                m_writer.Write(intValue);
         }
 
 
-        public void Write (string value) {
-            m_writer.Write(value);
+        public void Write (long longValue) {
+            m_writer.Write(longValue);
         }
 
 
-        public void Write (float value) {
-            m_writer.Write(value);
+        public void Write (long[] longs) {
+            m_writer.Write(longs.Length);
+            foreach (var longValue in longs)
+                m_writer.Write(longValue);
         }
 
 
-        public void Write (double value) {
-            m_writer.Write(value);
+        public void Write (char charValue) {
+            m_writer.Write(charValue);
         }
 
 
-        public void Write (bool value) {
-            m_writer.Write(value);
+        public void Write (char[] chars) {
+            m_writer.Write(chars.Length);
+            foreach (var charValue in chars)
+                m_writer.Write(charValue);
+        }
+
+
+        public void Write (string stringValue) {
+            m_writer.Write(stringValue);
+        }
+
+
+        public void Write (string[] strings) {
+            m_writer.Write(strings.Length);
+            foreach (var stringValue in strings)
+                m_writer.Write(stringValue);
+        }
+
+
+        public void Write (float floatValue) {
+            m_writer.Write(floatValue);
+        }
+
+
+        public void Write (float[] floats) {
+            m_writer.Write(floats.Length);
+            foreach (var floatValue in floats)
+                m_writer.Write(floatValue);
+        }
+
+
+        public void Write (double doubleValue) {
+            m_writer.Write(doubleValue);
+        }
+
+
+        public void Write (double[] doubles) {
+            m_writer.Write(doubles.Length);
+            foreach (var doubleValue in doubles)
+                m_writer.Write(doubleValue);
+        }
+
+
+        public void Write (bool boolValue) {
+            m_writer.Write(boolValue);
         }