+++ Tests/Runtime/TestMono.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-
-namespace SaveSystem.Tests.Runtime {
-
-    public class TestMono : MonoBehaviour { }
-
-}